Final Commit

+++ DeleteItemWindow.xaml.cs
@@ -1,18 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
+﻿using System.Configuration;
 using System.Data.SqlClient;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace PhoneBookAlpha
 {
@@ -29,7 +17,12 @@ namespace PhoneBookAlpha
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            string toDelete = nameToDelete.Text;
+            string toDelete = nameToDelete.Text.ToString().ToLower();
+            if (string.IsNullOrEmpty(toDelete))
+            {
+                MessageBox.Show("Please enter full name to remove it from Contacts");
+                return;
+            }
             string connectionString = ConfigurationManager.ConnectionStrings["PhoneBookAlpha.Properties.Settings.PhoneBookConnectionString"].ConnectionString;
             string query = "DELETE FROM PhoneBook WHERE Name = @DeleteIt";
             using (SqlConnection connection = new SqlConnection(connectionString))
@@ -46,8 +39,6 @@ namespace PhoneBookAlpha
             var startAgain = new MainWindow();
             startAgain.OpenConnection();
             startAgain.Show();
-
-
             Close();
         }
         public void CloseMainWindowNow()
